Move fields handling to a separate file
