+++ plugins/airtable/src/App.tsx
@@ -1,14 +1,15 @@
 import "./App.css"
 
 import type { ManagedCollection } from "framer-plugin"
+import type { DataSource } from "./data"
 
 import { framer } from "framer-plugin"
-import { useEffect, useLayoutEffect, useState } from "react"
-import { inferFields, type DataSource } from "./data"
-import { FieldMapping } from "./FieldMapping"
-import { SelectDataSource } from "./SelectDataSource"
 import { fetchTable } from "./api"
+import { inferFields } from "./fields"
+import { FieldMapping } from "./FieldMapping"
 import { NoTableAccess } from "./NoAccess"
+import { useEffect, useLayoutEffect, useState } from "react"
+import { SelectDataSource } from "./SelectDataSource"
 
 interface AppProps {
     collection: ManagedCollection
