+++ plugins/airtable/src/data.ts
@@ -5,10 +5,11 @@ import type {
     FieldDataEntryInput,
     FieldDataInput,
 } from "framer-plugin"
+import type { PossibleField } from "./fields"
 
 import { framer } from "framer-plugin"
 import { type AirtableFieldSchema, fetchAllBases, fetchRecords, fetchTable, fetchTables } from "./api"
-import { ALLOWED_FILE_TYPES, richTextToHTML } from "./utils"
+import { richTextToHTML } from "./utils"
 
 export const PLUGIN_KEYS = {
     BASE_ID: "airtablePluginBaseId",
@@ -45,265 +46,6 @@ export async function getTables(baseId: string, signal: AbortSignal): Promise<Ai
             fields: tableSchema.fields,
         }))
     )
-}
-
-type AllowedType = ManagedCollectionField["type"] | "unsupported"
-
-interface InferredField {
-    /**
-     * The type of the field as it appears in Airtable.
-     *
-     * Only set when fields are inferred.
-     */
-    readonly airtableType?: Exclude<AirtableFieldSchema["type"], "multipleRecordLinks" | "singleSelect">
-    readonly allowedTypes?: [AllowedType, ...AllowedType[]]
-}
-
-interface InferredMultipleRecordLinksField {
-    type: "collectionReference" | "multiCollectionReference"
-    readonly airtableType: "multipleRecordLinks"
-    readonly single: boolean
-    readonly supportedCollections: { id: string; name: string }[]
-    readonly allowedTypes?: []
-}
-
-interface InferredEnumField {
-    type: "enum"
-    readonly airtableType: "singleSelect"
-    readonly airtableCases: { id: string; name: string }[]
-    readonly allowedTypes: ["enum"]
-}
-
-export type PossibleField = (
-    | ManagedCollectionField
-    | {
-          id: string
-          name: string
-          userEditable: boolean
-          type: "unsupported"
-      }
-) &
-    (InferredField | InferredMultipleRecordLinksField | InferredEnumField)
-
-export async function inferFields(collection: ManagedCollection, table: AirtableTable): Promise<PossibleField[]> {
-    const fields: PossibleField[] = []
-
-    for (const fieldSchema of table.fields) {
-        const fieldMetadata = {
-            id: fieldSchema.id,
-            name: fieldSchema.name,
-            userEditable: false,
-        }
-
-        switch (fieldSchema.type) {
-            case "checkbox":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "boolean",
-                    allowedTypes: ["boolean"],
-                })
-                break
-
-            case "singleSelect":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "enum",
-                    cases: fieldSchema.options.choices.map(choice => ({
-                        id: choice.id,
-                        name: choice.name,
-                    })),
-                    airtableCases: fieldSchema.options.choices.map(choice => ({
-                        id: choice.id,
-                        name: choice.name,
-                    })),
-                    allowedTypes: ["enum"],
-                })
-                break
-
-            case "number":
-            case "percent":
-            case "currency":
-            case "autoNumber":
-            case "rating":
-            case "duration":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "number",
-                    allowedTypes: ["number"],
-                })
-                break
-
-            case "singleLineText":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "string",
-                    allowedTypes: ["string", "formattedText"],
-                })
-                break
-
-            case "email":
-            case "phoneNumber":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "string",
-                    allowedTypes: ["string", "link"],
-                })
-                break
-
-            case "multilineText":
-            case "richText":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "formattedText",
-                    allowedTypes: ["formattedText", "string"],
-                })
-                break
-
-            case "url":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "link",
-                    allowedTypes: ["link", "image", "file"],
-                })
-                break
-            case "multipleAttachments":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "image",
-                    allowedTypes: ["image", "file", "link"],
-                })
-                break
-
-            case "date":
-            case "dateTime":
-            case "createdTime":
-            case "lastModifiedTime":
-                fields.push({
-                    ...fieldMetadata,
-                    airtableType: fieldSchema.type,
-                    type: "date",
-                    allowedTypes: ["date"],
-                })
-                break
-
-            case "multipleRecordLinks": {
-                const foundCollections: { id: string; name: string }[] = []
-
-                if (table.id === fieldSchema.options.linkedTableId) {
-                    foundCollections.push({ id: collection.id, name: "This Collection" })
-                }
-
-                const existingCollections = await framer.getManagedCollections()
-                for (const existingCollection of existingCollections.filter(
-                    existingCollection => existingCollection.id !== collection.id
-                )) {
-                    const tableId = await existingCollection.getPluginData(PLUGIN_KEYS.TABLE_ID)
-
-                    if (tableId === fieldSchema.options.linkedTableId) {
-                        foundCollections.push({ id: existingCollection.id, name: existingCollection.name })
-                    }
-                }
-
-                const type = fieldSchema.options.prefersSingleRecordLink
-                    ? "collectionReference"
-                    : "multiCollectionReference"
-
-                // Only add unsupported field if we didn't find a matching collection
-                if (foundCollections.length === 0) {
-                    fields.push({
-                        ...fieldMetadata,
-                        airtableType: fieldSchema.type,
-                        type,
-                        collectionId: "",
-                        supportedCollections: [],
-                        single: fieldSchema.options.prefersSingleRecordLink,
-                    })
-                } else {
-                    fields.push({
-                        ...fieldMetadata,
-                        airtableType: "multipleRecordLinks",
-                        supportedCollections: foundCollections,
-                        collectionId: foundCollections[0].id,
-                        type,
-                        single: fieldSchema.options.prefersSingleRecordLink,
-                    })
-                }
-                break
-            }
-
-            case "formula": {
-                const result = fieldSchema.options.result
-                if (!result) {
-                    fields.push({
-                        ...fieldMetadata,
-                        type: "unsupported",
-                        allowedTypes: ["unsupported"],
-                    })
-                    continue
-                }
-
-                switch (result.type) {
-                    case "singleLineText":
-                    case "email":
-                    case "url":
-                    case "phoneNumber":
-                        fields.push({ ...fieldMetadata, type: "string", allowedTypes: ["string"] })
-                        break
-                    case "checkbox":
-                        fields.push({ ...fieldMetadata, type: "boolean", allowedTypes: ["boolean"] })
-                        break
-                    case "number":
-                    case "percent":
-                    case "currency":
-                    case "autoNumber":
-                    case "rating":
-                    case "duration":
-                        fields.push({ ...fieldMetadata, type: "number", allowedTypes: ["number"] })
-                        break
-                    case "multipleAttachments":
-                        fields.push({
-                            ...fieldMetadata,
-                            type: "file",
-                            allowedFileTypes: ALLOWED_FILE_TYPES,
-                            allowedTypes: ["file", "image", "link"],
-                        })
-                        break
-                    case "date":
-                    case "dateTime":
-                    case "createdTime":
-                    case "lastModifiedTime":
-                        fields.push({ ...fieldMetadata, type: "date", allowedTypes: ["date"] })
-                        break
-                    case "richText":
-                    case "multilineText":
-                        fields.push({
-                            ...fieldMetadata,
-                            type: "formattedText",
-                            allowedTypes: ["formattedText", "string"],
-                        })
-                        break
-                    default:
-                        fields.push({ ...fieldMetadata, type: "unsupported", allowedTypes: ["unsupported"] })
-                        continue
-                }
-                break
-            }
-
-            default:
-                fields.push({ ...fieldMetadata, type: "unsupported", allowedTypes: ["unsupported"] })
-                continue
-        }
-    }
-
-    return fields
 }
 
 const EMAIL_REGEX = /\S[^\s@]*@\S+\.\S+/
